requirements.txt
freetype-py         # For font loading
imageio             # For Image I/O
networkx            # For the scene graph
numpy               # Numpy
Pillow              # For Trimesh texture conversions
PyOpenGL            # For OpenGL
PyOpenGL_accelerate # For OpenGL
six                 # For Python 2/3 interop
trimesh             # For meshes
opencv-python
matplotlib
tqdm
shapely
h5py
pyglet == 1.4.0b1   # For the pyglet viewer